refactored contest ui
