+++ .gitignore
@@ -22,3 +22,6 @@ yarn-error.log*
 
 #TODO add .env to your gitignore
 
+# Eslint files
+.eslintrc.js
+.prettierrc.js
