+++ client/src/styles/styles.js
@@ -1,6 +1,20 @@
 import { makeStyles } from '@material-ui/core';
 
 const useStyles = makeStyles((theme) => ({
+  //nav bar
+  menuButton: {
+    marginRight: theme.spacing(2),
+  },
+
+  appBar: {
+    backgroundColor: theme.palette.primary.main,
+  },
+
+  toolBar: {
+    maxWidth: 900,
+    margin: 'auto',
+  },
+
   navLinks: {
     color: theme.palette.primary.light,
     marginRight: theme.spacing(3),
@@ -22,9 +36,16 @@ const useStyles = makeStyles((theme) => ({
     paddingTop: theme.spacing(6),
   },
 
+  container: {
+    width: '100%',
+  },
+
   innerContainer: {
     maxWidth: '75%',
-    height: '80rem',
+  },
+
+  titleLogo: {
+    flexGrow: 1,
   },
 
   title: {
@@ -34,27 +55,33 @@ const useStyles = makeStyles((theme) => ({
     marginBottom: theme.spacing(8),
     fontFamily: theme.typography.fontFamily[1],
   },
+
   subtitle: {
     textAlign: 'left',
     paddingTop: theme.spacing(6),
     fontFamily: theme.typography.fontFamily[1],
   },
+
   subtext: {
     color: theme.palette.grey.dark,
     paddingTop: theme.spacing(1),
     position: 'relative',
   },
+
   prizeTimeSub: {
     textAlign: 'left',
     fontFamily: theme.typography.fontFamily[1],
   },
+
   textFieldOne: {
     width: '100%',
     marginTop: theme.spacing(1),
   },
+
   paper: {
     backgroundColor: 'white',
   },
+
   grid: {
     display: 'flex',
     flexWrap: 'wrap',
@@ -62,16 +89,19 @@ const useStyles = makeStyles((theme) => ({
     overflow: 'hidden',
     paddingTop: theme.spacing(2),
   },
+
   gridList: {
     width: '100%',
     height: 400,
   },
+
   alignItemsAndJustifyContent: {
     width: '100%',
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'center',
   },
+
   contestButton: {
     height: 80,
     width: 240,
@@ -79,6 +109,7 @@ const useStyles = makeStyles((theme) => ({
     marginTop: theme.spacing(8),
     color: theme.palette.primary.light,
   },
+
   buttonText: {
     position: 'absolute',
     top: '50%',
